// Aysa_Risk_Front/src/app/core/services/models/IOptionParams..ts
import { HttpContext, HttpHeaders, HttpParams } from '@angular/common/http';

export interface OptionsParam {
  headers?:
    | HttpHeaders
    | {
        [header: string]: string | string[];
      };
  context?: HttpContext;
  observe?: 'body';
  params?:
    | HttpParams
    | {
        [param: string]: string | number | boolean | ReadonlyArray<string | number | boolean>;
      };
  reportProgress?: boolean;
  responseType?: 'json';
  withCredentials?: boolean;
}